% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_miranda.R
\name{get_miranda}
\alias{get_miranda}
\title{Prepare miRanda object}
\usage{
get_miranda(filename = NULL)
}
\arguments{
\item{filename}{The path to the miRanda download database in .txt format.
                If \code{NULL}, the database will be downloaded.
                Default: \code{NULL}}
}
\value{
A \code{GRanges} object corresponding to the mm10 miRanda database.
}
\description{
This function will download the mm10 miRanda database and convert it into
\code{GRanges} format.
}
\details{
Requires an internet connection unless a filename corresponding to the
downloaded database is provided.

By default, the "Good mirSVR score, Conserved miRNA" version of the database
is downloaded from the "August 2010 Release Downloads" section. See:
http://www.microrna.org/microrna/getDownloads.do
}
\examples{
\dontrun{
  miranda_gr <- miranda()
}
}

