% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/liftOver.R
\name{liftOver_mm9ToMm10}
\alias{liftOver_mm9ToMm10}
\title{Convert a GRanges from mm9 to mm10}
\usage{
liftOver_mm9ToMm10(gr)
}
\arguments{
\item{gr}{The \code{GRanges} to convert.}
}
\value{
A \code{GRanges} object corresponding to the mm10 positions.
}
\description{
Convert a GRanges from mm9 to mm10
}
\examples{
\dontrun{
  gr_mm10 <- liftOver_mm9ToMm10(gr_mm9)
}
}

