% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mm10.mirna.annotations-package.r
\name{miranda}
\alias{miranda}
\title{miRanda mm10}
\description{
This dataset is a \code{GRanges} object corresponding to the "Good mirSVR
score, Conserved miRNA" version of the database downloaded from the "August
2010 Release Downloads" section:
http://www.microrna.org/microrna/getDownloads.do
}
\details{
The data in the \code{GRanges} object correspond to the predicted miRNA
targets.

The \code{GRanges} have to following metadata columns:
  * mirna_name
  * gene_symbol

\itemize{
  \item mirna_name Name of the miRNA for the current target.
  \item gene_symbol Name of the targeted gene.
}
}

