% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miranda.R, R/mm10.mirna.annotations-package.r
\name{miranda}
\alias{miranda}
\title{Prepare miRanda object}
\usage{
miranda(filename = NULL)
}
\arguments{
\item{filename}{The path to the miRanda download database in .txt format.
                If \code{NULL}, the database will be downloaded.
                Default: \code{NULL}}
}
\value{
A \code{GRanges} object corresponding to the mm10 miRanda database.
}
\description{
This function will download the mm10 miRanda database and convert it into
\code{GRanges} format.

This dataset is a \code{GRanges} object corresponding to the "Good mirSVR
score, Conserved miRNA" version of the database downloaded from the "August
2010 Release Downloads" section:
http://www.microrna.org/microrna/getDownloads.do
}
\details{
Requires an internet connection unless a filename corresponding to the
downloaded database is provided.

By default, the "Good mirSVR score, Conserved miRNA" version of the database
is downloaded from the "August 2010 Release Downloads" section. See:
http://www.microrna.org/microrna/getDownloads.do

The data in the \code{GRanges} object correspond to the predicted miRNA
targets.

The \code{GRanges} have to following metadata columns:
  * mirna_name
  * gene_symbol

\itemize{
  \item mirna_name Name of the miRNA for the current target.
  \item gene_symbol Name of the targeted gene.
}
}
\examples{
\dontrun{
  miranda_gr <- miranda()
}
}

